reworked bundler
client bundler with the current asset combination
