+++ lib/client/bundler/asset.js
@@ -1,6 +1,3 @@
-// Client Asset File
-// -----------------
-// An asset is a Code (JS or CoffeeScript), CSS or HTML file
 'use strict';
 
 var formatKb, formatters, fs, jsp, log, minifyJSFile, pathlib, pro, uglifyjs, wrap, wrapCode,
@@ -12,11 +9,11 @@ pathlib = require('path');
 
 uglifyjs = require('uglify-js');
 
-formatters = require('./formatters');
+formatters = require('../formatters');
 
-log = require('../utils/log');
+log = require('../../utils/log');
 
-wrap = require('./wrap');
+wrap = require('../wrap');
 
 jsp = uglifyjs.parser;
 
@@ -25,8 +22,8 @@ pro = uglifyjs.uglify;
 // Load, compile and minify the following assets
 
 module.exports = function(ss, options) {
-  var loadFile;
-  loadFile = function(dir, fileName, type, options, cb) {
+
+  function loadFile(dir, fileName, type, options, cb) {
     var extension, formatter, path;
     dir = pathlib.join(ss.root, dir);
     path = pathlib.join(dir, fileName);
@@ -43,11 +40,10 @@ module.exports = function(ss, options) {
       throw new Error('Unable to render \'' + fileName + '\' as this appears to be a ' + (formatter.assetType.toUpperCase()) + ' file. Expecting some type of ' + (type.toUpperCase()) + ' file in ' + (dir.substr(ss.root.length)) + ' instead');
     }
     return formatter.compile(path.replace(/\\/g, '/'), options, cb);
-  };
+  }
+
+  
   return {
-
-    // Public
-
     js: function(path, opts, cb) {
       return loadFile(options.dirs.code, path, 'js', opts, function(output) {
         output = wrapCode(output, path, opts.pathPrefix, options);
