+++ lib/client/bundler/index.js
@@ -1,56 +1,41 @@
-function includeFlags(overrides) {
-  var includes = {
-    css: true,
-    html: true,
-    system: true,
-    initCode: true
-  };
-  if (overrides) {
-    for(var n in overrides) { includes[n] = overrides[n]; }
+
+var defaultBundler,
+    bundlers = {}; // get bundler by client name
+
+/**
+ * Define the bundler for a client
+ * @param client object to store the definition in
+ * @param args arguments passed to define
+ */
+exports.define = function bundler(ss,client,args,options) {
+
+  var name = args[0], 
+      pathsOrFunc = args[1];
+
+  defaultBundler = defaultBundler || require('./default')(ss,options);
+
+  if (typeof pathsOrFunc === "function") {
+    bundlers[name] = pathsOrFunc(ss,options);
+    bundlers[name].define(client, args[2], args[3], args[4], args[5]);
+  } else {
+    defaultBundler.define(client, args[1]);
   }
-  return includes;
-}
+};
 
+/**
+ * Determine the bundler for a client
+ * @param client Query params with client=name or an actual client object
+ */
+exports.get = function bundler(ss,client,options){
 
-module.exports = function bundler(ss,options){
-  var pack = require('./pack')(ss,options);
+  if (client.bundler) return client.bundler;
 
-  return {
+  if (typeof client.client === "string") {
+    return bundlers[client.client] || defaultBundler;
+  }
+  if (typeof client.name === "string") {
+    return bundlers[client.name] || defaultBundler;
+  }
 
-	  define: function define(client, paths) {
-
-      if (typeof paths.view !== 'string') {
-        throw new Error('You may only define one HTML view per single-page client. Please pass a filename as a string, not an Array');
-      }
-      if (paths.view.indexOf('.') === -1) {
-        throw new Error('The \'' + paths.view + '\' view must have a valid HTML extension (such as .html or .jade)');
-      }
-
-      // Alias 'templates' to 'tmpl'
-      if (paths.templates) {
-        paths.tmpl = paths.templates;
-      }
-
-      // Force each into an array
-      ['css', 'code', 'tmpl'].forEach(function(assetType) {
-        if (!(paths[assetType] instanceof Array)) {
-          paths[assetType] = [paths[assetType]];
-          return paths[assetType];
-        }
-      });
-
-      // Define new client object
-      client.paths = paths;
-      client.includes = includeFlags(paths.includes);
-
-      return client;
-  	},
-
-    packCSS: pack.css,
-    packJS: pack.js,
-
-    load: function load() {
-      //TODO something
-    }
-  };
+  return defaultBundler;
 };
