+++ lib/client/index.js
@@ -21,6 +21,7 @@ var options = {
   packedAssets: packAssets || false,
   liveReload: ['code', 'css', 'static', 'templates', 'views'],
   dirs: {
+    client: '/client',
     code: '/client/code',
     css: '/client/css',
     static: '/client/static',
