+++ lib/client/serve/ondemand.js
@@ -44,9 +44,11 @@ module.exports = function(ss, router, options) {
     output = [];
     dir = pathlib.join(ss.root, options.dirs.code);
     files = magicPath.files(dir, [path]);
+    //TODO determine client and get bundler
     return files.forEach(function(file) {
       var description;
       try {
+        //TODO return bundler.js()
         return asset.js(file, {
           pathPrefix: options.globalModules? null : path,
           compress: options.packAssets
