+++ lib/client/view.js
@@ -8,9 +8,8 @@ var pathlib = require('path'),
     wrap = require('./wrap');
 
 module.exports = function(ss, client, options, cb) {
-  var templateEngine = require('./template_engine')(ss);
-
-  var asset = require('./asset')(ss, options);
+  var templateEngine = require('./template_engine')(ss),
+      bundler = require('./bundler/index').get(ss, client, options);
 
   // Add links to CSS and JS files
   var includes = headers().concat(templates());
@@ -21,7 +20,7 @@ module.exports = function(ss, client, options, cb) {
     compress: options.packedAssets,
     filename: client.paths.view
   };
-  return asset.html(client.paths.view, htmlOptions, cb);
+  return bundler.asset.html(client.paths.client, htmlOptions, cb);
 
   // When packing assets the default path to the CSS or JS file can be overridden
   // either with a string or a function, typically pointing to an resource on a CDN
